requirements.txt
bottle
PyYAML
requests
requests_toolbelt
